Backend/pyproject.toml
[project]
name = "swiftbid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.53.0",
    "langchain>=1.1.2",
    "langchain-google-genai>=3.2.0",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
]
